fix(jobs): resolve the payload through a seam the test controls
The population guard `test_every_job_names_an_absolute_payload_not_a_bare_sac`
was red in CI while the behaviour it guards was correct on every host. It was
asserting a fact about the RUNNING ENVIRONMENT, not about the specs.

`sac_bin()` returns the bare name `"sac"` and warns when no console script sits
beside the running interpreter. That is correct for a broken install, and it is
also exactly what a PYTHONPATH-only CI run looks like: sac importable, console
script absent. So calling `provide_jobs()` with no seam and asserting the
payload is absolute asserts whichever environment happens to be running it.

THE FIX threads `sac_bin`'s existing `executable=` seam up through
`provide_jobs`, `accounts_jobs`, `maintenance_jobs` and `liveness_jobs`, so the
test resolves against a venv-shaped tree it builds on disk. Production keeps the
same behaviour: the fleet calls `provide_jobs()` with no argument and still gets
the console script beside the interpreter that imported the plugin.

WHY NOT THE OBVIOUS WEAKENING. "absolute OR a warning fired" would pass
vacuously for EVERY job in CI, because a warning always fires there -- including
for a job that hardcoded a literal `sac`. The guard would be green in the one
environment where it actually runs.

MUTATION-CHECKED with the mutant that shape deserves -- accept the seam, never
thread it down:

    3 failed, 221 passed
    killed by  test_every_payload_comes_from_the_tree_the_seam_named
               test_specs_built_without_a_console_script_warn
               test_specs_built_without_a_console_script_carry_the_bare_name
    NOT killed by  the absoluteness guard itself

The last line is the point. On a developer machine the default resolution is
absolute anyway, so the absoluteness guard passes for the wrong reason and
cannot tell an applied seam from an ignored one. That is what the new
`comes_from_the_tree` test exists for.

TWO REFACTORS THE PER-FILE CAP FORCED, both pure moves, both proved.

`_jobs_plugin.py` 538 -> 222 lines, split into `_specs_accounts.py` (the OAuth
credential beat, three jobs that share one failure mode) and
`_specs_maintenance.py` (drift, worktrees, images, freshness -- independent of
each other, alike in that every one REPORTS rather than repairs).
`_specs_liveness.py` already existed and says in its own docstring "Split out of
._jobs_plugin (at the per-file cap)", so this follows a precedent rather than
inventing a shape. Proof of pure move: the nine rendered JobSpecs -- name,
command, kind, schedule -- hash identically before and after
(sha256 b872c1f6dfbedbef).

`test__jobs_plugin.py` 847 -> 460 lines, split the same three ways, plus
`_jobspec_helpers.py` for the two readers the modules now share. Done by AST
rather than line range because the per-job tests are interleaved --
spartan_sif_bake at 491 and 500, restart_login_expired at 520 and 528,
spartan_sif_bake again at 539 -- so a contiguous cut would mix groups or drop
functions. Proof: the set of 224 collected test names hashes identically before
and after (2a0ef28621796390).

WHERE THE POPULATION GUARD LIVES, and why not in a file of its own. It was
briefly `test__specs_payload.py`, which the conformance audit correctly rejected
(PS-204 §2, orphan test file -- the repo mirrors tests onto source modules
one-to-one and there is no `_specs_payload.py`). It belongs in
`test__sac_bin.py`: its subject IS that module, whose entire docstring is the
argument for why a job command may not say bare `sac`. That file goes 96 -> 256
lines, well under the cap.

CI's lint scope was measured rather than assumed. `.github/workflows/lint.yml`
runs `ruff check --select F401,F811`, which is why the two pre-existing I001
findings in `test__jobs_plugin.py` do not fail the build and why the single F401
this change would otherwise have introduced -- `pathlib`, unused once the
population guard moved out -- would have. Only that one was fixed; the
pre-existing findings are left alone.

Verified: 224 passed in the `_jobs` tests, `tests/develop/test_audit.py` 2
passed in 72.78s (the duration is the evidence it ran -- a skipped audit
returns in hundredths), and `ruff check --select F401,F811` clean across
`src/_jobs/` and `tests/_jobs/`.

ONE PRE-EXISTING FAILURE FOUND AND NOT FIXED HERE. The whole suite under
`-n auto` reports
`test__agent_list_host_display.py::test_print_agent_list_started_column_renders_pinned_jst`
failing. Controls: it passes alone on develop (14 passed), passes alone in this
worktree (14 passed), and FAILS under `-n auto` on develop with none of these
changes present (1 failed, 3187 passed). So it is a parallel-execution
isolation defect that predates this branch. CI runs xdist too, so it is a
latent flake there rather than a dormant one.
