+++ tests/scitex_agent_container/_jobs/test__jobs_plugin.py
@@ -40,8 +40,6 @@ the provider import is lazy, so an old scitex-dev must not fail CI here.
 
 from __future__ import annotations
 
-import pathlib
-
 import re
 
 import pytest
@@ -53,30 +51,12 @@ jobs_mod = pytest.importorskip(
 
 from scitex_agent_container._jobs._jobs_plugin import provide_jobs  # noqa: E402
 
+from ._jobspec_helpers import _job, _split_command  # noqa: E402
 
 
-def _split_command(command: str) -> tuple[str, str, str]:
-    """``(bound, payload, rest)`` — the three things a job command can get wrong.
-
-    The whole-command assertions below used to pin a literal ``sac``. They
-    cannot any more: the payload is now the ABSOLUTE console script beside the
-    running interpreter (:mod:`._sac_bin`), so its text differs per machine and
-    a literal would pin the developer's venv into CI.
-
-    Splitting keeps every property those assertions were protecting — the bound
-    and the full argument list stay pinned character-for-character, so dropping
-    a ``--to`` or an ``--apply`` is still a red test — while the one token that
-    is legitimately machine-specific is checked by SHAPE instead. Asserting it
-    equals ``sac_bin()`` would be no test at all: a check parameterised by the
-    value it is checking cannot fail.
-    """
-    tokens = command.split()
-    return " ".join(tokens[:2]), tokens[2], " ".join(tokens[3:])
 
 
-def _job(name: str):
-    (match,) = [j for j in provide_jobs() if j.name == name]
-    return match
+
 
 
 
@@ -172,113 +152,22 @@ def test_provider_job_cadence_is_two_hours() -> None:
 # ---------------------------------------------------------------------------
 
 
-def test_accounts_keepalive_job_name_is_package_prefixed() -> None:
-    # Arrange — the distribution half of the single-refresher model.
-    # Act
-    job = _job("scitex-agent-container-accounts-keepalive")
-    # Assert
-    assert job.name == "scitex-agent-container-accounts-keepalive"
 
 
-def test_accounts_keepalive_job_kind_is_timer() -> None:
-    # Arrange — a periodic systemd --user timer, so kind="timer". A kind
-    # outside JobSpec.ALLOWED_KINDS raises at construction, and `ecosystem up`
-    # then silently DROPS sac's WHOLE provider (provider-isolated, WARN-only)
-    # — taking the OAuth refresh, the drift check, the worktree GC and both
-    # heal enforcers down with it.
-    # Act
-    job = _job("scitex-agent-container-accounts-keepalive")
-    # Assert
-    assert job.kind == "timer"
 
 
-def test_accounts_keepalive_command_pushes_to_every_access_only_peer() -> None:
-    # Arrange — the peer list IS the job. A keepalive that reaches two of the
-    # three access-only hosts looks healthy (exit 0, verified peers) while the
-    # third silently expires, which is the exact failure this job exists to
-    # end. Pin the whole command so dropping a `--to` is a red test, not a
-    # host nobody notices is dead.
-    # Act
-    job = _job("scitex-agent-container-accounts-keepalive")
-    # Assert
-    bound, _payload, rest = _split_command(job.command)
-    assert (bound, rest) == ("/usr/bin/timeout 300", "accounts keepalive --all --to ywata-note-win --to scitex-compute-03 --to scitex-compute-04")
 
 
-def test_accounts_keepalive_command_runs_the_copying_verb_not_a_minting_one() -> None:
-    # Arrange — belt-and-braces, the counterpart of accounts-refresh's
-    # --skip-active guard. This job COPIES the master's current token; minting
-    # rotates it, and a rotation revokes the token every running agent is
-    # holding. Scheduling `accounts refresh` or `accounts login` here would
-    # turn a keepalive into a fleet-wide logout every 15 minutes, so pin the
-    # verb itself rather than trusting the full-command assertion above to be
-    # re-read whenever someone edits the peer list.
-    #
-    # Indexed PAST the self-bounding `/usr/bin/timeout <N>` prefix (tokens 0
-    # and 1), so this still pins the VERB rather than the wrapper.
-    # Act
-    job = _job("scitex-agent-container-accounts-keepalive")
-    # Assert
-    _bound, _payload, rest = _split_command(job.command)
-    assert rest.split()[:2] == ["accounts", "keepalive"]
 
 
-def test_accounts_keepalive_cadence_bounds_the_follower_outage() -> None:
-    # Arrange — the cadence IS the worst-case follower outage, not a guess at
-    # when work is needed. The master's token changes once in ~7h at an
-    # unpredictable moment, and the instant it does every follower's copy is
-    # revoked; the tick only decides how long that revoked window lasts. A
-    # converged peer is verified rather than rewritten, so the extra ticks are
-    # near free — which is what makes 15min affordable.
-    # Act
-    job = _job("scitex-agent-container-accounts-keepalive")
-    # Assert
-    assert job.on_unit_active_sec == "15min"
 
 
-def test_accounts_keepalive_schedule_mirrors_the_timer_cadence() -> None:
-    # Arrange — the cron form is kept alongside the timer cadence (as every
-    # sibling does) so `ecosystem cron` could derive an equivalent line, and
-    # so the two spellings cannot silently disagree about how often this runs.
-    # Act
-    job = _job("scitex-agent-container-accounts-keepalive")
-    # Assert
-    assert job.schedule == "*/15 * * * *"
 
 
-def test_accounts_keepalive_timeout_outlives_a_three_peer_pass() -> None:
-    # Arrange — per peer: a handful of ssh ops plus ONE outbound HTTPS
-    # verification (15s cap inside the probe). 300s covers three peers
-    # including a slow one. A pass killed here is SAFE — nothing is published
-    # unverified, so the peer keeps its previous credential.
-    #
-    # The bound is asserted on the COMMAND, not on `timeout_sec`, because
-    # this job is deployed to cron and a cron line cannot carry that field.
-    # Act
-    job = _job("scitex-agent-container-accounts-keepalive")
-    # Assert
-    assert job.command.startswith("/usr/bin/timeout 300 ")
 
 
-def test_accounts_keepalive_constructs_as_a_real_jobspec() -> None:
-    # Arrange — construction must not raise (a bad field would drop the whole
-    # provider). Assert it is the canonical contract type, not a look-alike.
-    # Act
-    job = _job("scitex-agent-container-accounts-keepalive")
-    # Assert
-    assert isinstance(job, jobs_mod.JobSpec)
 
 
-def test_accounts_keepalive_and_accounts_refresh_are_both_declared() -> None:
-    # Arrange — unlike the heal pair below, these two are NOT alternatives:
-    # they are the two halves of one model and BOTH must be enabled. Refresh
-    # without keepalive leaves the followers holding a revoked copy; keepalive
-    # without refresh distributes a token nothing renews. Deleting either one
-    # breaks the fleet in a way that looks like the other half working.
-    # Act
-    names = {job.name for job in provide_jobs()}
-    # Assert
-    assert {"scitex-agent-container-accounts-refresh", "scitex-agent-container-accounts-keepalive"} <= names
 
 
 def test_provider_does_not_federate_listen_it_would_duplicate_the_supervisor() -> None:
@@ -305,205 +194,40 @@ def test_provider_does_not_federate_listen_it_would_duplicate_the_supervisor() -
     assert "sac.listen" not in names
 
 
-def test_host_sync_check_job_name_is_package_prefixed() -> None:
-    # Arrange — the drift-alarm timer that makes the Stage-0 detector run.
-    # Act
-    job = _job("scitex-agent-container-host-sync-check")
-    # Assert
-    assert job.name == "scitex-agent-container-host-sync-check"
 
 
-def test_host_sync_check_job_kind_is_timer() -> None:
-    # Arrange — a periodic systemd --user timer (hourly), so kind="timer".
-    # Act
-    job = _job("scitex-agent-container-host-sync-check")
-    # Assert
-    assert job.kind == "timer"
 
 
-def test_host_sync_check_command_is_the_readonly_check() -> None:
-    # Arrange — the scheduled command MUST carry --check. A timer that could
-    # fast-forward a peer unattended is Stage 1, explicitly out of scope.
-    # Act
-    job = _job("scitex-agent-container-host-sync-check")
-    # Assert
-    assert "--check" in job.command
 
 
-def test_host_sync_check_command_routes_to_a_seen_card() -> None:
-    # Arrange — --alarm is what turns the exit code into a SEEN board card
-    # instead of a journald line nobody reads.
-    # Act
-    job = _job("scitex-agent-container-host-sync-check")
-    # Assert
-    assert "--alarm" in job.command
 
 
-def test_host_sync_check_command_never_runs_the_mutating_remedy() -> None:
-    # Arrange — belt-and-braces: the exact mutating form `sac host sync
-    # <peer>` (no --check) must never be what this timer runs. The command
-    # is the read-only detector, full stop.
-    #
-    # `--exit-zero` was added 2026-08-17 and does NOT weaken that intent: it
-    # touches only this process's exit status, never a peer. It is here
-    # because the tri-state verdict (1 = drift found, 2 = undetermined) was
-    # being read by systemd as unit failure, which put the host into
-    # `degraded`, which made the dotfiles sync installer conclude systemd
-    # was absent and silently stop installing its timer.
-    #
-    # Kept as EXACT EQUALITY on purpose. This assertion is the reason that
-    # change could not land unnoticed — a substring check would have let it
-    # through silently, and the next edit to this command deserves the same
-    # scrutiny. Update the literal deliberately; do not relax the operator.
-    # Act
-    job = _job("scitex-agent-container-host-sync-check")
-    # Assert — the command is precisely the read-only check+alarm form.
-    bound, _payload, rest = _split_command(job.command)
-    assert (bound, rest) == ("/usr/bin/timeout 600", "host sync --check --all --alarm --exit-zero")
 
 
-def test_host_sync_check_cadence_is_hourly() -> None:
-    # Arrange — drift is slow-moving; hourly is ample and gentle on ssh.
-    # Act
-    job = _job("scitex-agent-container-host-sync-check")
-    # Assert
-    assert job.on_unit_active_sec == "1h"
 
 
-def test_worktree_gc_job_name_is_package_prefixed() -> None:
-    # Arrange — the daily GC that makes the worktree-sprawl countermeasure
-    # PERIODIC. A GC nobody schedules is a script, not a countermeasure —
-    # which is exactly how one repo reached 105 worktrees.
-    # Act
-    job = _job("scitex-agent-container-worktree-gc")
-    # Assert
-    assert job.name == "scitex-agent-container-worktree-gc"
 
 
-def test_worktree_gc_job_kind_is_timer() -> None:
-    # Arrange — a periodic systemd --user timer (daily), so kind="timer".
-    # A bad kind makes `ecosystem up` silently drop sac's WHOLE provider.
-    # Act
-    job = _job("scitex-agent-container-worktree-gc")
-    # Assert
-    assert job.kind == "timer"
 
 
-def test_worktree_gc_command_is_the_apply_form() -> None:
-    # Arrange — the scheduled job must ACT, not just report: a timer that
-    # only dry-runs would print a nightly report nobody reads while the
-    # sprawl kept growing. The safety lives in the predicate, not in
-    # withholding --apply.
-    # Act
-    job = _job("scitex-agent-container-worktree-gc")
-    # Assert
-    bound, _payload, rest = _split_command(job.command)
-    assert (bound, rest) == (
-        "/usr/bin/timeout 900",
-        # --exit-zero is LOAD-BEARING and pinned here on purpose. Exit 1 from
-        # this verb means "a repo is over cap, a human decides" — a finding,
-        # not ill health. Without the flag the supervisor records the job as
-        # failed, and the rule for retiring its duplicate systemd timer was
-        # "the supervisor has a recorded exit-0", which this job could then
-        # never produce. Dropping the flag silently re-creates that deadlock.
-        "worktree gc --apply --all --exit-zero",
-    )
 
 
-def test_worktree_gc_command_sweeps_every_declared_repo() -> None:
-    # Arrange — --all is only correct because it HAS a clean source (every
-    # agent spec.workdir that is a local git repo toplevel). If that source
-    # ever disappears, this command silently sweeps nothing.
-    # Act
-    job = _job("scitex-agent-container-worktree-gc")
-    # Assert
-    assert "--all" in job.command
 
 
-def test_worktree_gc_cadence_is_daily() -> None:
-    # Arrange — sprawl accumulates over days and the age gate is 24h, so a
-    # faster pass could not remove anything a daily one would miss.
-    # Act
-    job = _job("scitex-agent-container-worktree-gc")
-    # Assert
-    assert job.on_unit_active_sec == "1d"
 
 
-def test_fleet_reconcile_job_name_is_package_prefixed() -> None:
-    # Arrange — the enforcer of "should be running => is running".
-    # Act
-    job = _job("scitex-agent-container-fleet-reconcile")
-    # Assert
-    assert job.name == "scitex-agent-container-fleet-reconcile"
 
 
-def test_fleet_reconcile_job_kind_is_timer() -> None:
-    # Arrange — a periodic systemd --user timer, so kind="timer". A wrong
-    # kind raises at construction and `ecosystem up` then silently DROPS
-    # sac's whole provider (provider-isolated, WARN-only) — taking the OAuth
-    # refresh, the drift check and the worktree GC down with it.
-    # Act
-    job = _job("scitex-agent-container-fleet-reconcile")
-    # Assert
-    assert job.kind == "timer"
 
 
-def test_fleet_reconcile_command_is_the_applying_form() -> None:
-    # Arrange — THIS JOB IS THE MECHANISM. `restart.policy` in ~93 specs is
-    # dead code without it: `_lifecycle/_start.py` runs the loop that reads
-    # it on a daemon thread inside the short-lived `sac agents start` CLI, so
-    # the supervisor dies with the process that promised it. A scheduled
-    # DRY-RUN would restore nothing — the whole point is `--apply`.
-    # Act
-    job = _job("scitex-agent-container-fleet-reconcile")
-    # Assert
-    bound, _payload, rest = _split_command(job.command)
-    assert (bound, rest) == ("/usr/bin/timeout 300", "agents reconcile --apply")
 
 
-def test_fleet_reconcile_cadence_is_five_minutes() -> None:
-    # Arrange — the cadence IS the window a dead agent stays dead. A no-op
-    # pass is one batched `tmux list-sessions` plus a spec read each, so it
-    # is cheap enough to run often.
-    # Act
-    job = _job("scitex-agent-container-fleet-reconcile")
-    # Assert
-    assert job.on_unit_active_sec == "5min"
 
 
-def test_fleet_reconcile_timeout_outlives_a_capped_pass() -> None:
-    # Arrange — the pathological pass restarts `--limit` agents, each a
-    # stop+settle+start. A pass killed at this timeout is SAFE (the restart
-    # history is persisted per restart, not at the end), but the timeout must
-    # still comfortably exceed a normal pass or the enforcer never finishes.
-    #
-    # THIS is the job whose unbounded cron line was measured accumulating
-    # fourteen concurrent instances, the oldest 45 minutes old (2026-07-18).
-    # A `timeout_sec` assertion would have passed throughout that incident,
-    # because the field was set and the deployed cron line was still
-    # unbounded — so the bound is pinned where it actually runs.
-    # Act
-    job = _job("scitex-agent-container-fleet-reconcile")
-    # Assert
-    assert job.command.startswith("/usr/bin/timeout 300 ")
 
 
-def test_spartan_sif_bake_job_name_is_package_prefixed() -> None:
-    # Arrange — the daily remote SIF bake (operator directive 2026-07-17:
-    # bake on Spartan, rsync to the master, zero master CPU).
-    # Act
-    job = _job("scitex-agent-container-spartan-sif-bake")
-    # Assert
-    assert job.name == "scitex-agent-container-spartan-sif-bake"
 
 
-def test_spartan_sif_bake_job_kind_is_timer() -> None:
-    # Arrange — a periodic systemd --user timer (*/10), so kind="timer".
-    # A bad kind makes `ecosystem up` silently drop sac's WHOLE provider.
-    # Act
-    job = _job("scitex-agent-container-spartan-sif-bake")
-    # Assert
-    assert job.kind == "timer"
 
 
 # ---------------------------------------------------------------------------
@@ -517,95 +241,20 @@ def test_spartan_sif_bake_job_kind_is_timer() -> None:
 # ---------------------------------------------------------------------------
 
 
-def test_restart_login_expired_job_name_is_package_prefixed() -> None:
-    # Arrange — the auto-restarter for auth-dead-but-live agents.
-    # Act
-    job = _job("scitex-agent-container-restart-login-expired-agents")
-    # Assert
-    assert job.name == "scitex-agent-container-restart-login-expired-agents"
 
 
-def test_restart_login_expired_job_kind_is_timer() -> None:
-    # Arrange — a periodic systemd --user timer, so kind="timer". A wrong kind
-    # raises at construction and `ecosystem up` then silently DROPS sac's whole
-    # provider (provider-isolated, WARN-only) — taking the OAuth refresh, the
-    # drift check, the worktree GC AND the fleet-reconcile enforcer down too.
-    # Act
-    job = _job("scitex-agent-container-restart-login-expired-agents")
-    # Assert
-    assert job.kind == "timer"
 
 
-def test_spartan_sif_bake_command_is_the_confirmed_form() -> None:
-    # Arrange — `sac image bake-remote` REFUSES to run without --yes
-    # (exit 2), mirroring `sac image build`'s non-interactive gate. A
-    # scheduled command missing --yes would fail every single night —
-    # a timer that fires and does nothing, the inert-feature shape.
-    # Act
-    job = _job("scitex-agent-container-spartan-sif-bake")
-    # Assert
-    bound, _payload, rest = _split_command(job.command)
-    assert (bound, rest) == ("/usr/bin/timeout 14400", "image bake-remote --yes")
 
 
-def test_spartan_sif_bake_cadence_is_every_10_minutes() -> None:
-    # Arrange — the SIF is a point-in-time snapshot of @develop, and at our
-    # release rate a day-old one is mostly wrong. 30min was the operator's
-    # stated FLOOR (「最低でも30分に1回」), not his target — he asked why it was
-    # only every 30 and said even 1min would be fine — so this pins the
-    # cadence he wants. skip-if-unchanged keeps a no-change tick at one ssh
-    # round-trip instead of a multi-GB transfer, and the script's `flock -n`
-    # single-flights the overlaps a 10min interval necessarily causes.
-    # Act
-    job = _job("scitex-agent-container-spartan-sif-bake")
-    # Assert
-    assert job.on_unit_active_sec == "10min"
 
 
-def test_spartan_sif_bake_timeout_outlives_two_bakes_and_a_pull() -> None:
-    # Arrange — two full bakes (base + scitex) plus a multi-GB pull on a
-    # slow link must fit; the per-leg ssh timeout is 7200s, so the cap
-    # must exceed the worst legitimate chain or the timer kills its
-    # own successful runs.
-    #
-    # This job needs a REAL bound more than any sibling: at a */10 cadence
-    # an unbounded bake outlives its own tick by construction, so a wedged
-    # run piles up against every later one. The bound therefore has to be
-    # in the command — cron, the surface this is deployed to, cannot carry
-    # `timeout_sec` at all.
-    # Act
-    job = _job("scitex-agent-container-spartan-sif-bake")
-    # Assert
-    assert job.command.startswith("/usr/bin/timeout 14400 ")
 
 
-def test_restart_login_expired_command_is_the_applying_form() -> None:
-    # Arrange — a scheduled DRY-RUN would detect wedged agents and heal none.
-    # The whole point is `--apply`. Detection stays read-only; the restart is
-    # the only mutation.
-    # Act
-    job = _job("scitex-agent-container-restart-login-expired-agents")
-    # Assert
-    bound, _payload, rest = _split_command(job.command)
-    assert (bound, rest) == ("/usr/bin/timeout 300", "agents restart-login-expired --apply")
 
 
-def test_restart_login_expired_cadence_is_five_minutes() -> None:
-    # Arrange — the cadence IS the window a login-expired agent stays wedged,
-    # matched to fleet-reconcile so the two enforcers sweep on the same beat.
-    # Act
-    job = _job("scitex-agent-container-restart-login-expired-agents")
-    # Assert
-    assert job.on_unit_active_sec == "5min"
 
 
-def test_restart_login_expired_constructs_as_a_real_jobspec() -> None:
-    # Arrange — construction must not raise (a bad field would drop the whole
-    # provider). Assert it is the canonical contract type, not a look-alike.
-    # Act
-    job = _job("scitex-agent-container-restart-login-expired-agents")
-    # Assert
-    assert isinstance(job, jobs_mod.JobSpec)
 
 
 # ---------------------------------------------------------------------------
@@ -805,43 +454,7 @@ def test_the_cron_line_that_lands_on_the_host_carries_the_bound() -> None:
     assert all(" /usr/bin/timeout " in line for line in lines), lines
 
 
-def test_every_job_names_an_absolute_payload_not_a_bare_sac() -> None:
-    """The whole-fleet guard, and the one that would have caught compute-01.
-
-    Each assertion above pins ONE job, so a new JobSpec added tomorrow with a
-    bare ``sac`` would pass every one of them. This covers the population
-    instead: after the self-bounding ``/usr/bin/timeout`` head, the payload of
-    EVERY provided job must be an absolute path.
-
-    MEASURED 2026-08-20: with a relative payload, scitex-compute-01's ten sac
-    jobs sat at exit 127 with zero successes for hours, self-pull included, so
-    the host could not fetch the fix that would have repaired it. The other
-    three hosts were fine only because their supervisor inherited a PATH that
-    happened to contain the venv.
-    """
-    # Arrange
-    jobs = provide_jobs()
-    # Act
-    relative = [
-        (job.name, job.command)
-        for job in jobs
-        for payload in [_split_command(job.command)[1]]
-        if not pathlib.Path(payload).is_absolute()
-    ]
-    # Assert
-    assert relative == [], (
-        f"these commands name a payload that must be found on the ambient "
-        f"PATH: {relative}. `resolve_execstart` absolutises only the FIRST "
-        f"token, and that token is already `/usr/bin/timeout`, so nothing "
-        f"downstream will fix this."
-    )
-
-
-def test_the_payload_guard_has_something_to_guard() -> None:
-    """Non-vacuity for the check above — an empty provider would pass it silently."""
-    # Arrange
-    expected_minimum = 1
-    # Act
-    jobs = provide_jobs()
-    # Assert
-    assert len(jobs) >= expected_minimum
+# The POPULATION guard over these jobs' payloads — the one that catches a
+# JobSpec added tomorrow rather than any named above — lives in
+# `test__specs_payload.py`. It needs the `executable` seam to resolve
+# deterministically, which is a different setup from every test here.
